% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{all_ny_layers}
\alias{all_ny_layers}
\title{all_ny_layer2}
\usage{
all_ny_layers(net = NULL, k = 2:9 * 100, data_dir)
}
\arguments{
\item{net}{Weighted street network; loaded from `data_dir` if not provided}

\item{k}{Vector of widths of exponential decay (in m) for spatial interaction
models}

\item{data_dir}{The directory in which data are to be, or have previously
been, downloaded.}
}
\description{
Calculate all flow layer of pedestrian densities for New York City, for a
range of widths of exponential spatial interaction functions (`k`-values).
}
