% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-optimise.R
\name{calc_layer}
\alias{calc_layer}
\title{calc_layer
Calculate a single layer from specified origin and destination categories,
using specific values of 'k' and 'ks'.}
\usage{
calc_layer(net, from = "subway", to = "disperse", k, k_scale, data_dir)
}
\arguments{
\item{net}{Weighted street network; loaded from `data_dir` if not provided}

\item{from}{Category of origins for pedestrian flows; one of "subway" or
"residential"}

\item{to}{Category of destinations for pedestrian flows; one of
"residential", "education", "entertainment", "healthcare", "sustenance",
"transportation", or "disperse" for a general dispersal model.}

\item{k}{Width of exponential decay in metres}

\item{k_scale}{Scale `k` to size of origins (`s`), so
`k = k ^ (1 + s / smax)`.}

\item{data_dir}{The directory in which data are to be, or have previously
been, downloaded.}
}
\description{
calc_layer
Calculate a single layer from specified origin and destination categories,
using specific values of 'k' and 'ks'.
}
