% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-fit.R
\name{calc_layer_scales}
\alias{calc_layer_scales}
\title{calc_layer_scales}
\usage{
calc_layer_scales(data_dir, model = "final")
}
\arguments{
\item{data_dir}{The directory in which data are to be, or have previously
been, downloaded.}

\item{model}{Default "final" uses the final model of layers reduced only to
those with significance less than `sig_model` in \link{build_ped_model},
otherwise the full model with potentially less significant layers is used.}
}
\value{
A `data.frame` with six columns of (1) 'model' with the layer names;
(2) 'estimates' holding the estimates for each layer from the multiple linear
regression model; (3) 'full_rel' with the relative scaling coefficient for
the contribution of each layer weighted by the full flow layers; (4)
'ped_rel' with equivalent values weighted by values at pedestrian count
stations only; (5) 'full_abs' with the absolute scaling coefficient for each
layer weighted by full flow layers; and (6) 'ped_abs' with equivalent values
weighted by values at pedestrian count stations only.
}
\description{
Calculate relative scales of contributions of each flow layer to final
model of pedestrian flows generated by successive calls to
\link{build_ped_model}.
}
