% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{centrality_edge_to_point}
\alias{centrality_edge_to_point}
\title{centrality_edge_to_point}
\usage{
centrality_edge_to_point(data_dir, save = FALSE)
}
\arguments{
\item{data_dir}{The data directory}

\item{save}{If `TRUE`, save resultant vertex-based centrality measures in
`data_dir`.}
}
\value{
Equivalent point- or vertex-based measures of centrality
}
\description{
Convert edge-based centrality measures to equivalent point-based.
}
