% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-download.R
\name{clb_osm_data}
\alias{clb_osm_data}
\title{clb_osm_data}
\usage{
clb_osm_data(city, prefix, data_dir = NULL)
}
\arguments{
\item{city}{City or bounding box for which data are to be extracted}

\item{prefix}{Prefix to prepend to file names to identify city}

\item{data_dir}{Directory in which data are to be stored}
}
\value{
List of OSM data for highways, green spaces, and activity attractors.
}
\description{
Get OpenStreetMap calibration data.
}
