% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-to-ped-pts.R
\name{fit_cen_to_ped}
\alias{fit_cen_to_ped}
\title{fit_cen_to_ped}
\usage{
fit_cen_to_ped(data_dir)
}
\arguments{
\item{data_dir}{The directory in which data are to be, or have previously
been, downloaded.}
}
\value{
A list containing vectors of 'k' and 'n' values, and a matrix of
30x20 = 600 columns, one for each combination of 30 'k'- and 20 'n'-values.
}
\description{
Fit network centrality to pedestrian counts, by aggregating across a variable
number of edges ('n') nearest to each pedestrian count station.
}
