% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowlayers.R
\name{fit_flows_to_ped}
\alias{fit_flows_to_ped}
\title{fit_flows_to_ped}
\usage{
fit_flows_to_ped(net_f, data_dir)
}
\arguments{
\item{net_f}{A network with flow columns, obtained from `get_layer`}

\item{data_dir}{The directory in which data are to be, or have previously
been, downloaded.}
}
\value{
A list containing vectors of 'k' and 'n' values, and a matrix of
30x20 = 600 columns, one for each combination of 30 'k'- and 20 'n'-values.
}
\description{
Fit one network with multiple flow columns (from multiple k-values) to
pedestrian counts, by aggregating each flow column across a variable number
of edges ('n') nearest to each pedestrian count station.
}
