% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-cache.R
\name{fit_one_layer}
\alias{fit_one_layer}
\title{fit_one_layer}
\usage{
fit_one_layer(net = NULL, from = "subway", to = "activity",
  k = 700, k_scale = 0, data_dir, cache = FALSE, quiet = FALSE)
}
\value{
Vector containing R-squared statistic and sum of squared errors for
fitted mode (divided by 1e6)
}
\description{
Calculate fit for one layer with observed pedestrian counts, and return both
sum of squared errors, and R-squared value
}
