% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-optimise.R
\name{get_layer}
\alias{get_layer}
\title{get_layer}
\usage{
get_layer(net, from = "subway", to = "disperse", data_dir)
}
\arguments{
\item{net}{Weighted street network; loaded from `data_dir` if not provided}

\item{from}{Category of origins for pedestrian flows; one of "subway" or
"residential"}

\item{to}{Category of destinations for pedestrian flows; one of
"residential", "education", "entertainment", "healthcare", "sustenance",
"transportation", or "disperse" for a general dispersal model.}

\item{data_dir}{The directory in which data are to be, or have previously
been, downloaded.}
}
\description{
Append network with flow columns between nominated places
}
