% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{ny_layer}
\alias{ny_layer}
\title{ny_layer}
\usage{
ny_layer(net = NULL, from = "subway", to = "activity", k = 700,
  k_scale = 0, data_dir, quiet = FALSE)
}
\arguments{
\item{net}{Weighted street network; loaded from `data_dir` if not provided}

\item{from}{Category of origins for pedestrian flows; one of "subway" or
"residential"}

\item{to}{Category of destinations for pedestrian flows; one of
"residential", "education", "entertainment", "healthcare", "sustenance",
"transportation", or "disperse" for a general dispersal model.}

\item{k}{Width of exponential decay (in m) for spatial interaction models}

\item{k_scale}{Scale `k` to size of origins (`s`), so `k = k ^ (1 + s /
smax)`.}

\item{data_dir}{The directory in which data are to be, or have previously
been, downloaded.}

\item{quiet}{If `FALSE`, display progress information on screen}
}
\description{
Calculate one flow layer of pedestrian densities for New York City
}
