% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyosm.R
\name{nyosm_data}
\alias{nyosm_data}
\title{nyosm_data}
\usage{
nyosm_data(data_dir)
}
\arguments{
\item{data_dir}{The directory in which data are to be, or have previously
been, downloaded.}
}
\value{
A list of three items: (1) the street network; (2) building polygons;
and (3) green space polygons
}
\description{
Get OpenStreetMap data for New York City
}
