% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyped-package.R
\docType{package}
\name{nyped-package}
\alias{nyped}
\alias{nyped-package}
\title{nyped: Model of pedestrian flows throughout New York City}
\description{
Model of pedestrian flows throughout New York City.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ATFutures/nyped}
  \item Report bugs at \url{https://github.com/ATFutures/nyped/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com} (\href{https://orcid.org/0000-0003-2172-5265}{ORCID})

}
\keyword{internal}
