% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyped.R
\name{nyped_data}
\alias{nyped_data}
\title{nyped_data}
\usage{
nyped_data(data_dir = tempdir(), quiet = FALSE)
}
\arguments{
\item{data_dir}{The directory in which data are to be, or have previously
been, downloaded.}

\item{quiet}{If `FALSE`, display progress information on screen}
}
\value{
A `data.frame` of pedestrian counts, and geographical coordinates,
with counts for weekdays, weekends, and "week" derived as a weighted
combintaion of both.
}
\description{
Download, clean, and return New York City pedestrian count data
}
\examples{
dat <- nyped_data ()
# library (mapview)
# mapview (dat, cex = "week", zcol = "week")
}
