% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nypopdens.R
\name{nypopdens}
\alias{nypopdens}
\title{nypopdens}
\usage{
nypopdens(data_dir = tempdir())
}
\arguments{
\item{data_dir}{The directory in which data are to be, or have previously
been, downloaded.}
}
\value{
Population density data
}
\description{
Get population density data for New York City
}
