% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nysubway.R
\name{nysubway}
\alias{nysubway}
\title{nysubway}
\usage{
nysubway()
}
\value{
A `data.frame` of subway naes, counts, and geographical coordinates
}
\description{
Get New York City subway station counts and coordinates
}
