% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-optimise.R
\name{optim_layer1}
\alias{optim_layer1}
\title{optim_layer1}
\usage{
optim_layer1(net, from = "subway", to = "disperse", flowvars = NULL,
  data_dir, cache = TRUE)
}
\arguments{
\item{net}{Weighted street network; loaded from `data_dir` if not provided}

\item{from}{Category of origins for pedestrian flows; one of "subway" or
"residential"}

\item{to}{Category of destinations for pedestrian flows; one of
"residential", "education", "entertainment", "healthcare", "sustenance",
"transportation", or "disperse" for a general dispersal model.}

\item{flowvars}{A `data.frame` of flows from previously calculated layers. If
provided, these are included in the model, and the target layer is optimised
as part of a multiple linear regression which includes these variables. Must
have same number of rows as `net`.}

\item{data_dir}{The directory in which data are to be, or have previously
been, downloaded.}

\item{cache}{If `TRUE`, layers are cached in a sub-directory of `data_dir`
for later reloading.}
}
\description{
Optimise flow layer to observed pedestrian densities according to single `k`
parameter. Note that optimisation can't be implemented directly because
results are too noisy, so a series of custom ranges is scanned, and optimal
values found from loess fits.
}
