% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-optimise.R
\name{optim_layer2}
\alias{optim_layer2}
\title{optim_layer2}
\usage{
optim_layer2(net, from = "subway", to = "disperse", k = NULL,
  data_dir)
}
\arguments{
\item{net}{Weighted street network; loaded from `data_dir` if not provided}

\item{from}{Category of origins for pedestrian flows; one of "subway" or
"residential"}

\item{to}{Category of destinations for pedestrian flows; one of
"residential", "education", "entertainment", "healthcare", "sustenance",
"transportation", or "disperse" for a general dispersal model.}

\item{k}{Width of exponential decay (in m) for spatial interaction models}

\item{data_dir}{The directory in which data are to be, or have previously
been, downloaded.}
}
\description{
Optimise `k` and `k_scale` parameters to optimally fit flow layer to observed
pedestrian densities. Note that optimisation can't be implemented directly
because results are too noisy, so a series of custom ranges is scanned, and
optimal values found from loess fits.
}
