% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-optimise.R
\name{optimise_layer}
\alias{optimise_layer}
\title{optimise_layer}
\usage{
optimise_layer(net, from, to, data_dir = data_dir)
}
\arguments{
\item{net}{Weighted street network; loaded from `data_dir` if not provided}

\item{from}{Category of origins for pedestrian flows; one of "subway" or
"residential"}

\item{to}{Category of destinations for pedestrian flows; one of
"residential", "education", "entertainment", "healthcare", "sustenance",
"transportation", or "disperse" for a general dispersal model.}

\item{data_dir}{The directory in which data are to be, or have previously
been, downloaded.}
}
\description{
Optimise `k` and `k_scale` parameters to optimally fit flow layer to observed
pedestrian densities
}
