% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-fit.R
\name{ped_model_to_full_flow}
\alias{ped_model_to_full_flow}
\title{ped_model_to_full_flow}
\usage{
ped_model_to_full_flow(mod, data_dir)
}
\arguments{
\item{mod}{A final pedestrian model produced by successive calls to
\link{build_ped_model}}

\item{data_dir}{The directory in which data are to be, or have previously
been, downloaded.}
}
\description{
Convert a final model of pedestrian flows at the pedestrian count stations
back into a model of full flows along each edge of the entire network
}
