% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_ny_popdens}
\alias{plot_ny_popdens}
\title{plot_ny_popdens}
\usage{
plot_ny_popdens(data_dir)
}
\arguments{
\item{data_dir}{The directory in which data are to be, or have previously
been, downloaded.}
}
\description{
Plot the population density of New York City via 'mapdeck'
}
\note{
This function presumes that a mapdeck token exists as an environmental
variable with a name that includes "mapbox"
}
