% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_ped_counts}
\alias{plot_ped_counts}
\title{plot_ped_counts}
\usage{
plot_ped_counts(type = "week")
}
\arguments{
\item{type}{One of "weekday", "weekend", or "week" (aggregate of both)}
}
\description{
Plot pedestrian count data for New York City
}
