% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_subway_counts}
\alias{plot_subway_counts}
\title{plot_subway_counts}
\usage{
plot_subway_counts(year = 2018L)
}
\arguments{
\item{year}{Year in [2013:2018]}
}
\description{
Plot subway entrace/exit count data for New York City
}
