% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowlayers.R
\name{all_flows_to_ped}
\alias{all_flows_to_ped}
\title{all_flows_to_ped}
\usage{
all_flows_to_ped(data_dir, sub_exits = TRUE)
}
\arguments{
\item{data_dir}{The directory in which data are to be, or have previously
been, downloaded.}

\item{sub_exits}{(For flows to/from subway only.) Calculate layer from subway
exits (`TRUE`), or just from single points denoting subway stations
(`FALSE`)?}
}
\description{
Batch-convert all flow layers to equivalent values matched to pedestrian
counting stations via the \link{fit_flows_to_ped} function.
}
