% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-optimise.R
\name{build_ped_model}
\alias{build_ped_model}
\title{build_ped_model}
\usage{
build_ped_model(data_dir, dat = NULL, sig = 0.01)
}
\arguments{
\item{data_dir}{The directory in which data are to be, or have previously
been, downloaded.}

\item{dat}{Output of previous run of `build_ped_model`}

\item{sig}{Desired level of statistical significance}
}
\description{
Build final pedestrian model
}
