% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration-package.R
\docType{package}
\name{calibration-package}
\alias{calibration}
\alias{calibration-package}
\title{calibration: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/moveability/calibration}
  \item Report bugs at \url{https://github.com/moveability/calibration/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com} (0000-0003-2172-5265)

}
\keyword{internal}
