% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-fns.R
\name{cut_network_to_pts}
\alias{cut_network_to_pts}
\title{cut_network_to_pts}
\usage{
cut_network_to_pts(data_dir, k = 700)
}
\arguments{
\item{data_dir}{The directory in which data are to be, or have previously
been, downloaded.}

\item{k}{Exponential decay parameter to be used in spatial interaction
models. This deterines the radius at which to cut the network.}
}
\description{
Cut the New York City street network into portions surrounding each
pedestrian counting point.
}
