% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{fit_one_layer}
\alias{fit_one_layer}
\title{fit_one_layer}
\usage{
fit_one_layer(net = NULL, from = "subway", to = "activity",
  k = 700, k_scale = 0, data_dir, cache = FALSE, quiet = FALSE)
}
\arguments{
\item{net}{Weighted street network; loaded from `data_dir` if not provided}

\item{from}{Category of origins for pedestrian flows; one of "subway" or
"residential"}

\item{to}{Category of destinations for pedestrian flows; one of
"residential", "education", "entertainment", "healthcare", "sustenance",
"transportation", or "disperse" for a general dispersal model.}

\item{k}{Width of exponential decay (in m) for spatial interaction models}

\item{k_scale}{Scale `k` to size of origins (`s`), so `k = k ^ (1 + s /
smax)`.}

\item{data_dir}{The directory in which data are to be, or have previously
been, downloaded.}

\item{cache}{If `TRUE`, layers are cached in a sub-directory of `data_dir`
for later reloading.}

\item{quiet}{If `FALSE`, display progress information on screen}
}
\value{
Vector containing R-squared statistic and sum of squared errors for
fitted mode (divided by 1e6)
}
\description{
Calculate fit for one layer with observed pedestrian counts, and return both
sum of squared errors, and R-squared value
}
