% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{ny_layer}
\alias{ny_layer}
\title{ny_layer}
\usage{
ny_layer(net = NULL, from = "subway", to = "activity", k = 700,
  data_dir)
}
\arguments{
\item{net}{Weighted street network; loaded from `data_dir` if not provided}

\item{from}{Category of origins for pedestrian flows}

\item{to}{Category of destinations for pedestrian flows; one of "activity",
"parking", or anything else to estimate dispersal.}

\item{k}{Width of exponential decay (in m) for spatial interaction models}

\item{data_dir}{The directory in which data are to be, or have previously
been, downloaded.}
}
\description{
Calculate one flow layer of pedestrian densities for New York City
}
