% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nysubway.R
\name{nysubway_data}
\alias{nysubway_data}
\title{nysubway_data}
\usage{
nysubway_data(quiet = FALSE)
}
\arguments{
\item{quiet}{If `FALSE`, display progress information on screen}
}
\value{
A `data.frame` of subway names, annual counts, and geographical
coordinates.
}
\description{
Download, clean, and join New York City subway station counts and coordinates
}
\examples{
dat <- nysubway_data ()
# library (mapview)
# mapview (dat, cex = "count2018", zcol = "count2018")
}
